feat: pattern for feature/files
