+++ frontend/src/app/global.css
@@ -116,4 +116,3 @@
 .custom-scrollbar::-webkit-scrollbar-corner {
   @apply bg-transparent;
 }
-
