+++ frontend/src/app/layout.tsx
@@ -14,7 +14,6 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      
       <body className=' overflow-hidden'><TitleBar />
       {children}
       </body>
