+++ frontend/src/entities/tabs/model/store.ts
@@ -1,5 +1,3 @@
-// entities/tab/model/tab.store.ts
-
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
