+++ frontend/src/features/file/create/model/useCreateFile.ts
@@ -1,0 +1,7 @@
+
+export const useCreateFileModel = () => {
+
+    return {
+        
+    }
+}
