+++ frontend/src/features/file/delete/model/useDeleteFileModel.ts
@@ -1,0 +1,7 @@
+
+export const useDeleteFile = () => {
+
+    return {
+        
+    }
+}
