+++ frontend/src/features/file/rename/model/useRenameFile.ts
@@ -1,0 +1,7 @@
+
+export const useRenameFile = () => {
+
+    return {
+        
+    }
+}
