+++ frontend/src/features/file/save/model/useSaveFile.ts
@@ -1,0 +1,7 @@
+
+export const useSaveFile = () => {
+
+    return {
+
+    }
+}
